# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * OnFocus multi-vendor discovery platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ListingType = typeof ListingType[keyof typeof ListingType];


export const ListingType = {
  artist: 'artist',
  vendor: 'vendor',
  venue: 'venue',
} as const;

export interface Listing {
  id: number;
  type: ListingType;
  name: string;
  category: string;
  city: string;
  rating: number;
  reviewCount: number;
  bio: string;
  coverImage: string;
  images: string[];
  tags: string[];
  verified: boolean;
  featured: boolean;
  /** @nullable */
  yearsActive?: number | null;
  /** @nullable */
  eventsCompleted?: number | null;
  /** @nullable */
  priceRange?: string | null;
  /** @nullable */
  capacity?: number | null;
  /** @nullable */
  amenities?: string[] | null;
}

export interface ListingsPage {
  listings: Listing[];
  total: number;
  limit: number;
  offset: number;
}

export type ListingInputType = typeof ListingInputType[keyof typeof ListingInputType];


export const ListingInputType = {
  artist: 'artist',
  vendor: 'vendor',
  venue: 'venue',
} as const;

export interface ListingInput {
  type: ListingInputType;
  name: string;
  category: string;
  city: string;
  bio: string;
  tags?: string[];
  priceRange?: string;
  capacity?: number;
}

export interface FeaturedListings {
  artists: Listing[];
  vendors: Listing[];
  venues: Listing[];
}

export interface PlatformStats {
  totalProfessionals: number;
  eventsSupported: number;
  citiesCovered: number;
  verifiedProfiles: number;
}

export interface CategoryItem {
  slug: string;
  label: string;
  count: number;
}

export interface CategoriesGrouped {
  artists: CategoryItem[];
  vendors: CategoryItem[];
  venues: CategoryItem[];
}

export interface InquiryInput {
  name: string;
  email: string;
  message: string;
  subject?: string;
}

export type PartnerApplicationInputType = typeof PartnerApplicationInputType[keyof typeof PartnerApplicationInputType];


export const PartnerApplicationInputType = {
  artist: 'artist',
  vendor: 'vendor',
  venue: 'venue',
} as const;

export interface PartnerApplicationInput {
  name: string;
  email: string;
  phone?: string;
  type: PartnerApplicationInputType;
  category: string;
  city: string;
  description: string;
  website?: string;
}

export interface InquiryConfirmation {
  success: boolean;
  message: string;
}

export type ListListingsParams = {
type?: ListListingsType;
category?: string;
city?: string;
search?: string;
featured?: boolean;
limit?: number;
offset?: number;
};

export type ListListingsType = typeof ListListingsType[keyof typeof ListListingsType];


export const ListListingsType = {
  artist: 'artist',
  vendor: 'vendor',
  venue: 'venue',
} as const;