# lib/api-zod/src/generated/types/categoriesGrouped.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * OnFocus multi-vendor discovery platform API
 * OpenAPI spec version: 0.1.0
 */
import type { CategoryItem } from './categoryItem';

export interface CategoriesGrouped {
  artists: CategoryItem[];
  vendors: CategoryItem[];
  venues: CategoryItem[];
}