# lib/api-zod/src/generated/types/inquiryConfirmation.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * OnFocus multi-vendor discovery platform API
 * OpenAPI spec version: 0.1.0
 */

export interface InquiryConfirmation {
  success: boolean;
  message: string;
}