# lib/api-zod/src/generated/types/listing.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * OnFocus multi-vendor discovery platform API
 * OpenAPI spec version: 0.1.0
 */
import type { ListingType } from './listingType';

export interface Listing {
  id: number;
  type: ListingType;
  name: string;
  category: string;
  city: string;
  rating: number;
  reviewCount: number;
  bio: string;
  coverImage: string;
  images: string[];
  tags: string[];
  verified: boolean;
  featured: boolean;
  /** @nullable */
  yearsActive?: number | null;
  /** @nullable */
  eventsCompleted?: number | null;
  /** @nullable */
  priceRange?: string | null;
  /** @nullable */
  capacity?: number | null;
  /** @nullable */
  amenities?: string[] | null;
}