# lib/api-zod/src/generated/types/listingInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * OnFocus multi-vendor discovery platform API
 * OpenAPI spec version: 0.1.0
 */
import type { ListingInputType } from './listingInputType';

export interface ListingInput {
  type: ListingInputType;
  name: string;
  category: string;
  city: string;
  bio: string;
  tags?: string[];
  priceRange?: string;
  capacity?: number;
}