# lib/api-zod/src/generated/types/listingsPage.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * OnFocus multi-vendor discovery platform API
 * OpenAPI spec version: 0.1.0
 */
import type { Listing } from './listing';

export interface ListingsPage {
  listings: Listing[];
  total: number;
  limit: number;
  offset: number;
}