# lib/api-zod/src/generated/types/partnerApplicationInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * OnFocus multi-vendor discovery platform API
 * OpenAPI spec version: 0.1.0
 */
import type { PartnerApplicationInputType } from './partnerApplicationInputType';

export interface PartnerApplicationInput {
  name: string;
  email: string;
  phone?: string;
  type: PartnerApplicationInputType;
  category: string;
  city: string;
  description: string;
  website?: string;
}