# lib/api-zod/src/generated/types/platformStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * OnFocus multi-vendor discovery platform API
 * OpenAPI spec version: 0.1.0
 */

export interface PlatformStats {
  totalProfessionals: number;
  eventsSupported: number;
  citiesCovered: number;
  verifiedProfiles: number;
}